pyproject.toml
[tool.poetry]
name = "aws pipeline stack"
version = "0.1.0"
description = "Pipeline using a diverse AWS services and dbt"
authors = ["Emiliano Frigo"]

package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0.0"
black = { extras = ["jupyter"], version = "^24.10.0" }
jupyter-black = "^0.4.0"
boto3 = "^1.24.48"
ipykernel = "^6.6.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
